Move TalkerViewAppBar in widgets folder
